Se hace ajuste para leer todas las facturas Se agrega seeder para usuario Se agrega pantalla de carga

+++ database/migrations/2021_11_16_170003_create_invoices_table.php
@@ -16,9 +16,9 @@ class CreateInvoicesTable extends Migration
         Schema::create('invoices', function (Blueprint $table) {
             $table->id();
             $table->string('id_siigo');
-            $table->integer('number');
+            $table->bigInteger('number');
             $table->string('name');
-            $table->integer('total');
+            $table->bigInteger('total');
             $table->string('status');
             $table->string('id_service')->nullable();
             $table->dateTime('date_service')->nullable();
